bigquery argument setter config code refactoring
